Add versioning to the C++ portion of the library.
