+++ tests/CMakeLists.txt
@@ -53,6 +53,7 @@ set(TEST_SOURCES
     ${TEST_DIR}/svs/lib/threads/thread.cpp
     ${TEST_DIR}/svs/lib/threads/threadlocal.cpp
     ${TEST_DIR}/svs/lib/threads/threadpool.cpp
+    ${TEST_DIR}/svs/lib/version.cpp
     ${TEST_DIR}/svs/lib/uuid.cpp
     # Third Party
     ${TEST_DIR}/svs/third-party/fmt.cpp
